Refactored Find User by Username Command and Request
