+++ src/api-authentication/src/test/java/org/rbernalop/apiauthentication/user/application/finder/username/FindUserByUsernameQueryHandlerTest.java
@@ -3,7 +3,9 @@ package org.rbernalop.apiauthentication.user.application.finder.username;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
-import org.rbernalop.apiauthentication.user.application.create.UserMother;
+import org.rbernalop.apiauthentication.shared.application.userfinder.username.FindUserByUsernameQuery;
+import org.rbernalop.apiauthentication.shared.application.userfinder.username.FindUserByUsernameResponse;
+import org.rbernalop.apiauthentication.user.domain.aggregate.UserMother;
 import org.rbernalop.apiauthentication.user.domain.aggregate.User;
 import org.rbernalop.apiauthentication.user.domain.port.UserRepository;
 import org.rbernalop.apiauthentication.user.domain.value_object.UserUsername;
