+++ src/api-authentication/src/test/java/org/rbernalop/apiauthentication/user/domain/aggregate/UserMother.java
@@ -1,6 +1,7 @@
-package org.rbernalop.apiauthentication.user.application.create;
+package org.rbernalop.apiauthentication.user.domain.aggregate;
 
-import org.rbernalop.apiauthentication.user.domain.aggregate.User;
+import org.rbernalop.apiauthentication.user.application.create.CreateUserCommand;
+import org.rbernalop.apiauthentication.user.application.create.CreateUserCommandMother;
 import org.rbernalop.apiauthentication.user.domain.value_object.*;
 
 public class UserMother {
